Factor out integration tests
