+++ qubic-rpc/Cargo.toml
@@ -3,6 +3,11 @@ name = "qubic-rpc"
 version = "0.1.0"
 edition = "2021"
 
+[[bin]]
+name = "server"
+test = false
+bench = false
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
